Refactor SqlServer provider into partial modules
