+++ DbaClientX.SqlServer/SqlServer.BulkOperations.cs
@@ -1,0 +1,202 @@
+using System;
+using System.Data;
+using System.Threading;
+using System.Threading.Tasks;
+using Microsoft.Data.SqlClient;
+
+namespace DBAClientX;
+
+public partial class SqlServer
+{
+    /// <summary>
+    /// Uses <see cref="SqlBulkCopy"/> to insert data contained in <paramref name="table"/> into the specified destination table.
+    /// </summary>
+    public virtual void BulkInsert(
+        string serverOrInstance,
+        string database,
+        bool integratedSecurity,
+        DataTable table,
+        string destinationTable,
+        bool useTransaction = false,
+        int? batchSize = null,
+        int? bulkCopyTimeout = null,
+        string? username = null,
+        string? password = null)
+    {
+        if (table == null)
+        {
+            throw new ArgumentNullException(nameof(table));
+        }
+
+        var connectionString = BuildConnectionString(serverOrInstance, database, integratedSecurity, username, password);
+
+        SqlConnection? connection = null;
+        var dispose = false;
+        if (useTransaction)
+        {
+            if (_transaction == null || _transactionConnection == null)
+            {
+                throw new DbaTransactionException("Transaction has not been started.");
+            }
+
+            connection = _transactionConnection;
+        }
+        else
+        {
+            connection = CreateConnection(connectionString);
+            OpenConnection(connection);
+            dispose = true;
+        }
+
+        try
+        {
+            using var bulkCopy = CreateBulkCopy(connection!, useTransaction ? _transaction : null);
+            bulkCopy.DestinationTableName = destinationTable;
+            if (batchSize.HasValue)
+            {
+                bulkCopy.BatchSize = batchSize.Value;
+            }
+            if (bulkCopyTimeout.HasValue)
+            {
+                bulkCopy.BulkCopyTimeout = bulkCopyTimeout.Value;
+            }
+
+            foreach (DataColumn column in table.Columns)
+            {
+                bulkCopy.ColumnMappings.Add(column.ColumnName, column.ColumnName);
+            }
+
+            WriteToServer(bulkCopy, table);
+        }
+        catch (Exception ex)
+        {
+            throw new DbaQueryExecutionException("Failed to execute bulk insert.", destinationTable, ex);
+        }
+        finally
+        {
+            if (dispose)
+            {
+                connection?.Dispose();
+            }
+        }
+    }
+
+    /// <summary>
+    /// Asynchronously uses <see cref="SqlBulkCopy"/> to insert data contained in <paramref name="table"/> into the specified destination table.
+    /// </summary>
+    public virtual async Task BulkInsertAsync(
+        string serverOrInstance,
+        string database,
+        bool integratedSecurity,
+        DataTable table,
+        string destinationTable,
+        bool useTransaction = false,
+        int? batchSize = null,
+        int? bulkCopyTimeout = null,
+        CancellationToken cancellationToken = default,
+        string? username = null,
+        string? password = null)
+    {
+        if (table == null)
+        {
+            throw new ArgumentNullException(nameof(table));
+        }
+
+        var connectionString = BuildConnectionString(serverOrInstance, database, integratedSecurity, username, password);
+
+        SqlConnection? connection = null;
+        var dispose = false;
+        if (useTransaction)
+        {
+            if (_transaction == null || _transactionConnection == null)
+            {
+                throw new DbaTransactionException("Transaction has not been started.");
+            }
+
+            connection = _transactionConnection;
+        }
+        else
+        {
+            connection = CreateConnection(connectionString);
+            await OpenConnectionAsync(connection, cancellationToken).ConfigureAwait(false);
+            dispose = true;
+        }
+
+        try
+        {
+            using var bulkCopy = CreateBulkCopy(connection!, useTransaction ? _transaction : null);
+            bulkCopy.DestinationTableName = destinationTable;
+            if (batchSize.HasValue)
+            {
+                bulkCopy.BatchSize = batchSize.Value;
+            }
+            if (bulkCopyTimeout.HasValue)
+            {
+                bulkCopy.BulkCopyTimeout = bulkCopyTimeout.Value;
+            }
+
+            foreach (DataColumn column in table.Columns)
+            {
+                bulkCopy.ColumnMappings.Add(column.ColumnName, column.ColumnName);
+            }
+
+            await WriteToServerAsync(bulkCopy, table, cancellationToken).ConfigureAwait(false);
+        }
+        catch (Exception ex)
+        {
+            throw new DbaQueryExecutionException("Failed to execute bulk insert.", destinationTable, ex);
+        }
+        finally
+        {
+            if (dispose)
+            {
+                connection?.Dispose();
+            }
+        }
+    }
+
+    /// <summary>
+    /// Creates a configured <see cref="SqlBulkCopy"/> instance for the supplied connection and transaction.
+    /// </summary>
+    /// <param name="connection">An open SQL Server connection.</param>
+    /// <param name="transaction">Optional transaction to enlist the bulk copy operation in.</param>
+    /// <returns>A configured <see cref="SqlBulkCopy"/> instance.</returns>
+    protected virtual SqlBulkCopy CreateBulkCopy(SqlConnection connection, SqlTransaction? transaction) => new(connection, SqlBulkCopyOptions.Default, transaction);
+
+    /// <summary>
+    /// Writes the contents of <paramref name="table"/> to the server using the provided bulk copy instance.
+    /// </summary>
+    /// <param name="bulkCopy">The configured bulk copy instance.</param>
+    /// <param name="table">Source data table.</param>
+    protected virtual void WriteToServer(SqlBulkCopy bulkCopy, DataTable table) => bulkCopy.WriteToServer(table);
+
+    /// <summary>
+    /// Asynchronously writes the contents of <paramref name="table"/> to the server using the provided bulk copy instance.
+    /// </summary>
+    /// <param name="bulkCopy">The configured bulk copy instance.</param>
+    /// <param name="table">Source data table.</param>
+    /// <param name="cancellationToken">Token used to cancel the bulk copy operation.</param>
+    /// <returns>A task that completes when the transfer finishes.</returns>
+    protected virtual Task WriteToServerAsync(SqlBulkCopy bulkCopy, DataTable table, CancellationToken cancellationToken) => bulkCopy.WriteToServerAsync(table, cancellationToken);
+
+    /// <summary>
+    /// Creates a new <see cref="SqlConnection"/> for the supplied connection string.
+    /// </summary>
+    /// <param name="connectionString">Connection string used to construct the connection.</param>
+    /// <returns>An unopened <see cref="SqlConnection"/> instance.</returns>
+    protected virtual SqlConnection CreateConnection(string connectionString) => new(connectionString);
+
+    /// <summary>
+    /// Opens the provided <see cref="SqlConnection"/>.
+    /// </summary>
+    /// <param name="connection">Connection to open.</param>
+    protected virtual void OpenConnection(SqlConnection connection) => connection.Open();
+
+    /// <summary>
+    /// Asynchronously opens the provided <see cref="SqlConnection"/>.
+    /// </summary>
+    /// <param name="connection">Connection to open.</param>
+    /// <param name="cancellationToken">Token used to cancel the operation.</param>
+    /// <returns>A task that completes when the connection is open.</returns>
+    protected virtual Task OpenConnectionAsync(SqlConnection connection, CancellationToken cancellationToken) => connection.OpenAsync(cancellationToken);
+}
