+++ DbaClientX.SqlServer/SqlServer.Streaming.cs
@@ -1,0 +1,140 @@
+#if NETSTANDARD2_1_OR_GREATER || NETCOREAPP3_0_OR_GREATER
+using System.Collections.Generic;
+using System.Data;
+using System.Data.Common;
+using System.Runtime.CompilerServices;
+using System.Threading;
+using System.Threading.Tasks;
+using Microsoft.Data.SqlClient;
+
+namespace DBAClientX;
+
+public partial class SqlServer
+{
+    /// <summary>
+    /// Asynchronously executes a SQL query and streams the resulting rows.
+    /// </summary>
+    public virtual IAsyncEnumerable<DataRow> QueryStreamAsync(
+        string serverOrInstance,
+        string database,
+        bool integratedSecurity,
+        string query,
+        IDictionary<string, object?>? parameters = null,
+        bool useTransaction = false,
+        [EnumeratorCancellation] CancellationToken cancellationToken = default,
+        IDictionary<string, SqlDbType>? parameterTypes = null,
+        IDictionary<string, ParameterDirection>? parameterDirections = null,
+        string? username = null,
+        string? password = null)
+    {
+        return Stream();
+
+        async IAsyncEnumerable<DataRow> Stream()
+        {
+            var connectionString = BuildConnectionString(serverOrInstance, database, integratedSecurity, username, password);
+
+            SqlConnection? connection = null;
+            var dispose = false;
+            try
+            {
+                (connection, dispose) = await ResolveConnectionAsync(connectionString, useTransaction, cancellationToken).ConfigureAwait(false);
+                var dbTypes = ConvertParameterTypes(parameterTypes);
+                await foreach (var row in ExecuteQueryStreamAsync(connection, useTransaction ? _transaction : null, query, parameters, cancellationToken, dbTypes, parameterDirections).ConfigureAwait(false))
+                {
+                    yield return row;
+                }
+            }
+            finally
+            {
+                if (dispose)
+                {
+                    connection?.Dispose();
+                }
+            }
+        }
+    }
+
+    /// <summary>
+    /// Asynchronously executes a stored procedure and streams the resulting rows.
+    /// </summary>
+    public virtual IAsyncEnumerable<DataRow> ExecuteStoredProcedureStreamAsync(
+        string serverOrInstance,
+        string database,
+        bool integratedSecurity,
+        string procedure,
+        IDictionary<string, object?>? parameters = null,
+        bool useTransaction = false,
+        [EnumeratorCancellation] CancellationToken cancellationToken = default,
+        IDictionary<string, SqlDbType>? parameterTypes = null,
+        IDictionary<string, ParameterDirection>? parameterDirections = null,
+        string? username = null,
+        string? password = null)
+    {
+        return Stream();
+
+        async IAsyncEnumerable<DataRow> Stream()
+        {
+            var connectionString = BuildConnectionString(serverOrInstance, database, integratedSecurity, username, password);
+
+            SqlConnection? connection = null;
+            var dispose = false;
+            try
+            {
+                (connection, dispose) = await ResolveConnectionAsync(connectionString, useTransaction, cancellationToken).ConfigureAwait(false);
+                var dbTypes = ConvertParameterTypes(parameterTypes);
+                await foreach (var row in ExecuteQueryStreamAsync(connection, useTransaction ? _transaction : null, procedure, parameters, cancellationToken, dbTypes, parameterDirections, commandType: CommandType.StoredProcedure).ConfigureAwait(false))
+                {
+                    yield return row;
+                }
+            }
+            finally
+            {
+                if (dispose)
+                {
+                    connection?.Dispose();
+                }
+            }
+        }
+    }
+
+    /// <summary>
+    /// Asynchronously executes a stored procedure using <see cref="DbParameter"/> instances and streams the resulting rows.
+    /// </summary>
+    public virtual IAsyncEnumerable<DataRow> ExecuteStoredProcedureStreamAsync(
+        string serverOrInstance,
+        string database,
+        bool integratedSecurity,
+        string procedure,
+        IEnumerable<DbParameter>? parameters = null,
+        bool useTransaction = false,
+        [EnumeratorCancellation] CancellationToken cancellationToken = default,
+        string? username = null,
+        string? password = null)
+    {
+        return Stream();
+
+        async IAsyncEnumerable<DataRow> Stream()
+        {
+            var connectionString = BuildConnectionString(serverOrInstance, database, integratedSecurity, username, password);
+
+            SqlConnection? connection = null;
+            var dispose = false;
+            try
+            {
+                (connection, dispose) = await ResolveConnectionAsync(connectionString, useTransaction, cancellationToken).ConfigureAwait(false);
+                await foreach (var row in ExecuteQueryStreamAsync(connection, useTransaction ? _transaction : null, procedure, cancellationToken: cancellationToken, dbParameters: parameters, commandType: CommandType.StoredProcedure).ConfigureAwait(false))
+                {
+                    yield return row;
+                }
+            }
+            finally
+            {
+                if (dispose)
+                {
+                    connection?.Dispose();
+                }
+            }
+        }
+    }
+}
+#endif
